backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.8.2
httpx==0.27.0
requests==2.32.3
pandas==2.2.2
numpy==1.26.4
scipy==1.13.1
python-dotenv==1.0.1
